requirements.txt
fastapi==0.101.1
uvicorn==0.23.2
sqlalchemy==2.0.20
psycopg2==2.9.7
psycopg2-binary==2.9.7
alembic==1.12.0
httpx==0.24.1
python-jose==3.3.0
passlib==1.7.4
pydantic==2.3.0
python-dotenv==1.0.0
authlib==1.2.1
itsdangerous==2.1.2
pydantic-settings
pydantic[email]
starlette-sessions
asyncpg